conv split into black/white and rgb
