+++ apps/coreir_examples/Makefile
@@ -1,4 +1,4 @@
-ALL_COREIR_APPS = pointwise conv
+ALL_COREIR_APPS = pointwise conv_bw conv_rgb
 
 cleanall:
 	for app in $(ALL_COREIR_APPS); do \
