--- apps/coreir_examples/conv/Makefile
@@ -1,50 +1,0 @@
-#### Halide flags
-HALIDE_BIN_PATH := ../../..
-HALIDE_SRC_PATH := ../../..
-include ../../support/Makefile.inc
-
-#### HLS flags
-include ../hls_support/Makefile.inc
-HLS_LOG = vivado_hls.log
-
-#### COREIR flags
-COREIR_DIR ?= ../../../../coreir
-COREIR_LIBPATH = $(COREIR_DIR)/build/coreir.a
-
-.PHONY: all run_hls
-all: out.png
-run_hls: $(HLS_LOG)
-
-
-pipeline: pipeline.cpp
-	$(CXX) $(CXXFLAGS) -Wall -g $^ $(LIB_HALIDE) $(COREIR_LIBPATH) -o $@ $(LDFLAGS) -ltinfo
-
-pipeline_hls.cpp pipeline_native.o pipeline_cuda.o: pipeline
-	HL_DEBUG_CODEGEN=0 ./pipeline
-
-run: run.cpp pipeline_hls.cpp hls_target.cpp pipeline_native.o
-	$(CXX) $(CXXFLAGS) -O1 -DNDEBUG $(HLS_CXXFLAGS) -g -Wall -Werror $^ -lpthread -ldl -o $@
-
-
-run_cuda: pipeline_native.o pipeline_cuda.o run_cuda.cpp
-	$(CXX) -O3 $(CXXFLAGS) -Wall -Werror $^ -lpthread -ldl -o $@  $(PNGFLAGS) `pkg-config --libs opencv`
-
-out.png: run
-	./run
-
-$(HLS_LOG): ../hls_support/run_hls.tcl pipeline_hls.cpp run.cpp
-	RUN_PATH=$(realpath ./) \
-	RUN_ARGS= \
-	vivado_hls -f $< -l $(HLS_LOG)
-
-out_cuda.png: run_cuda
-	CUDA_LAUNCH_BLOCKING=1 HL_NUM_THREADS=4 ./run_cuda ../../images/benchmark_1080p_rgb.png
-
-clean:
-	rm -f pipeline run
-	rm -f *.html
-	rm -f pipeline_native.h pipeline_native.o
-	rm -f pipeline_hls.h pipeline_hls.cpp
-	rm -f hls_target.h hls_target.cpp
-	rm -f pipeline_coreir.cpp pipeline_coreir.h
-	rm -f coreir_target.h coreir_target.cpp design_top.json
