--- apps/coreir_examples/conv/design_target.json
@@ -1,195 +1,0 @@
-{
-  "namespaces": {
-    "_G": {
-      "modules": {
-        "DesignTarget": {
-          "def": null,
-          "type": [
-            "Record",
-            [
-              [
-                "in",
-                [
-                  "Array",
-                  16,
-                  "BitIn"
-                ]
-              ],
-              [
-                "out",
-                [
-                  "Array",
-                  16,
-                  "BitOut"
-                ]
-              ]
-            ]
-          ]
-        }
-      }
-    },
-    "stdlib": {
-      "generators": {
-        "add2": {
-          "genparams": {
-            "width": "int"
-          },
-          "typegen": [
-            "stdlib",
-            "binop"
-          ]
-        }
-      },
-      "modules": {
-        "GPI_16": {
-          "type": [
-            "Record",
-            [
-              [
-                "out",
-                [
-                  "Array",
-                  16,
-                  "BitOut"
-                ]
-              ]
-            ]
-          ]
-        },
-        "GPO_16": {
-          "type": [
-            "Record",
-            [
-              [
-                "in",
-                [
-                  "Array",
-                  16,
-                  "BitIn"
-                ]
-              ]
-            ]
-          ]
-        },
-        "add2_16": {
-          "type": [
-            "Record",
-            [
-              [
-                "in0",
-                [
-                  "Array",
-                  16,
-                  "BitIn"
-                ]
-              ],
-              [
-                "in1",
-                [
-                  "Array",
-                  16,
-                  "BitIn"
-                ]
-              ],
-              [
-                "out",
-                [
-                  "Array",
-                  16,
-                  "BitOut"
-                ]
-              ]
-            ]
-          ]
-        },
-        "const_16": {
-          "configparams": {
-            "value": "int"
-          },
-          "type": [
-            "Record",
-            [
-              [
-                "out",
-                [
-                  "Array",
-                  16,
-                  "BitOut"
-                ]
-              ]
-            ]
-          ]
-        },
-        "mult2_16": {
-          "type": [
-            "Record",
-            [
-              [
-                "in0",
-                [
-                  "Array",
-                  16,
-                  "BitIn"
-                ]
-              ],
-              [
-                "in1",
-                [
-                  "Array",
-                  16,
-                  "BitIn"
-                ]
-              ],
-              [
-                "out",
-                [
-                  "Array",
-                  16,
-                  "BitOut"
-                ]
-              ]
-            ]
-          ]
-        }
-      },
-      "namedtypegens": [
-        {
-          "genparams": {
-            "width": "int"
-          },
-          "name": "binop"
-        },
-        {
-          "flippedname": "uintIn",
-          "genparams": {
-            "width": "int"
-          },
-          "name": "uint"
-        },
-        {
-          "flippedname": "intIn",
-          "genparams": {
-            "width": "int"
-          },
-          "name": "int"
-        }
-      ],
-      "namedtypes": [
-        {
-          "flippedname": "rstIn",
-          "name": "rst",
-          "rawtype": "BitOut"
-        },
-        {
-          "flippedname": "clkIn",
-          "name": "clk",
-          "rawtype": "BitOut"
-        }
-      ]
-    }
-  },
-  "top": [
-    "_G",
-    "DesignTarget"
-  ]
-}
